+++ bevy_rapier2d/examples/custom_system_setup2.rs
@@ -1,4 +1,4 @@
-use bevy::{core::FrameCount, prelude::*, transform::TransformSystem};
+use bevy::{diagnostic::FrameCount, prelude::*, transform::TransformSystem};
 use bevy_rapier2d::prelude::*;
 
 fn main() {
@@ -61,7 +61,7 @@ fn despawn_one_box(
 }
 
 fn setup_graphics(mut commands: Commands) {
-    commands.spawn((Camera2d::default(), Transform::from_xyz(0.0, 20.0, 0.0)));
+    commands.spawn((Camera2d, Transform::from_xyz(0.0, 20.0, 0.0)));
 }
 
 pub fn setup_physics(mut commands: Commands) {