+++ bevy_rapier3d/examples/debugdump3.rs
@@ -1,16 +1,16 @@
 //! Example using bevy_mod_debugdump to output a graph of systems execution order.
 //! run with:
-//! `cargo run --example debugdump2 > dump.dot && dot -Tsvg dump.dot > dump.svg`
+//! `cargo run --example debugdump3 > dump.dot && dot -Tsvg dump.dot > dump.svg`
 
 use bevy::prelude::*;
 use bevy_mod_debugdump::{schedule_graph, schedule_graph_dot};
-use bevy_rapier2d::prelude::*;
+use bevy_rapier3d::prelude::*;
 
 fn main() {
     let mut app = App::new();
     app.add_plugins((
         DefaultPlugins,
-        RapierPhysicsPlugin::<NoUserData>::pixels_per_meter(100.0),
+        RapierPhysicsPlugin::<NoUserData>::default(),
         RapierDebugRenderPlugin::default(),
     ));