+++ bevy_rapier3d/examples/joints3.rs
@@ -281,17 +281,16 @@ pub fn setup_physics(mut commands: Commands) {
 pub fn print_impulse_revolute_joints(
     context: ReadRapierContext,
     joints: Query<(Entity, &ImpulseJoint)>,
-) {
+) -> Result<()> {
     for (entity, impulse_joint) in joints.iter() {
-        match &impulse_joint.data {
-            TypedJoint::RevoluteJoint(_revolute_joint) => {
-                println!(
-                    "angle for {}: {:?}",
-                    entity,
-                    context.single().impulse_revolute_joint_angle(entity),
-                );
-            }
-            _ => {}
+        if let TypedJoint::RevoluteJoint(_revolute_joint) = &impulse_joint.data {
+            println!(
+                "angle for {}: {:?}",
+                entity,
+                context.single()?.impulse_revolute_joint_angle(entity),
+            );
         }
     }
+
+    Ok(())
 }